requirements.txt
Django==4.2.16
psycopg2-binary==2.9.6
djangorestframework==3.15.2
requests==2.32.3
django-extensions==3.2.1
factory_boy==3.2.1
python-dotenv==1.0.1